restructure cli setup for poetry
